refactored project. Created a feast_finder module with the logic and updated the app.py. Also, removed the Favorite_Recipe class
