+++ .gitignore
@@ -2,3 +2,4 @@ back-end/.env
 */.env
 */__pycache__/*
 back-end/tests/__pycache__
+back-end/feast_finder/__pycache__
